+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Change Log
 
-All notable changes to the "translator-helper" extension will be documented in this file.
+## [0.3.0] - 2022-05-09
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Automatically copy the translated text to clipboard for the `translate` command.
+- Support selection hover translation. You can select a word or a sentence, then you will see a popup with the translated text. Click **Copy** to copy the translated text. You can turn off this feature in the settings.
 
-## [Unreleased]
+## [0.2.0] - 2022-05-08
 
-- Initial release
+- Update dependencies.
+
+## [0.1.0]
+
+- Add reverse translation.
+
+## [0.0.3]
+
+- Initial release.